+++ src/mame/sinclair/atm.cpp
@@ -14,8 +14,7 @@ NOTES:
 #include "atm.h"
 
 #include "bus/ata/atapicdr.h"
-#include "bus/ata/hdd.h"
-#include "sound/ay8910.h"
+#include "bus/ata/idehd.h"
 
 #define LOG_MEM   (1U << 1)
 #define LOG_VIDEO (1U << 2)
@@ -75,7 +74,7 @@ void atm_state::atm_update_memory()
 
 u16 atm_state::atm_update_memory_get_page(u8 bank)
 {
-	return m_pen ? pen_page(bank) : (u16) (~PEN_RAMNROM_MASK & ~PEN_DOS7FFD_MASK);
+	return m_pen ? pen_page(bank) : (~PEN_RAMNROM_MASK & ~PEN_DOS7FFD_MASK);
 }
 
 void atm_state::atm_ula_w(offs_t offset, u8 data)
